% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eva surfs.R
\name{eva_surfs}
\alias{eva_surfs}
\title{A function to report back on Eva's surfing}
\usage{
eva_surfs(total_waves, wave_size, day)
}
\arguments{
\item{total_waves}{Total number of waves Eva caught, should be a positive number}

\item{wave_size}{Approx. wave size per day}

\item{day}{day of the surfing}
}
\description{
A function to report back on Eva's surfing
}
